Reduces the self-purge effect from Haloperidol (#89361)
## About The Pull Request

This reverts my first ever PR which made Haloperidol purge itself while
players were in stamina crit. This change was initially made because
this chem (believe it or not, it's not the only chem that was capable of
this) could stack in exceedingly high amounts and keep players in
stamina crit for MINUTES on end. It REALLY wasn't fun for those who got
subjected to it during my experimentation with it.

However, after the PR
(https://github.com/tgstation/tgstation/pull/83445) which went under my
radar (this was made nearly 2 months after my first pr). Getting
permanently stamina crit is no longer possible, especially via chems. So
the Haloperidol self-purge (being the band-aid it was) is no longer
necessary.

No, I'm not doing this because i want to buff haloperidol. With or
without the self-purge, ya'll really need to learn what calomel is.


EDIT: Maintainer assigned to this is unhappy and won't let this be
merged without compromise.
Title changed from "Removes the self-purge effect from Haloperidol"
And the self-purge is instead now reduced to double it's existing
metabolization speed, rather than 2 units it's effectively: 0.16 + 0.16
(while stam crit)

## Why It's Good For The Game

Removes some unnecessary band-aid code. (While maintaining balance?)

## Changelog

:cl:
balance: Haloperidol's self-purge is drastically reduced while the host
is stamina crit, as a result of permanent stamina crit no longer being a
thing to worry about.
/:cl:

---------

Co-authored-by: Aki Ito <11748095+ExcessiveUseOfCobblestone@users.noreply.github.com>