Fixes Cryogelidia trapping you in stasis hell even though the cube broke (#89396)
## About The Pull Request

Rather than mimic the effects of the chilled extract's cube, we actually
use the status effect to accomplish that goal, except it does not give
you the option to resist out of the cube. On exposure, our mob is cubed
along with being put into stasis.

If the status effect is removed (because someone broke the cube), we
purge the chem from our system.

When we are no longer metabolizing the chemical, we also remove the cube
and the stasis effect.

## Why It's Good For The Game

Someone asked me to fix it, and I'm obliging. Hopefully this comparable
to prior behavior. I make no promises. but it appears to be working as
expected.

## Changelog
:cl:
fix: Cryogelidia can no longer put you into stasis longer than the
lifespan of the cube prison it kept you in.
/:cl: