Fixes runtime spam from `end_reaction` (#89368)
## About The Pull Request

It seems as if in some contexts, a reagent holder will end up with no
associated atom, but with ongoing chemical reactions

It will attempt to stop these reactions, either naturally or by being
forced to (from the aforementioned event of disassociating with its
atom), but this will runtime due to this has trait check.

I actually have a feeling this proc was originally intended to be pure
but uhh I dunno for sure

I'm not *100%* sure if this is the correct fix or if it's silencing a
runtime. Originally I was gonna add a stack trace but I decided against
it - If any reviewer would prefer that I'll re-add it