fixed the space dragon overlay assuming the base color (#89433)
## About The Pull Request
For some reason the unit test screenshot was unaffected but the overlay
of the space dragon shouldn't inherit the color (unless the dragon is
hit with colorful reagent or w/e) of the atom. This is most noticeably
on the space shark form which has an underside that should be white,
which is learly not happening right now.

## Why It's Good For The Game
Fixing a small issue I have noticed while watching moocows videos on
conveyor belts shenanigan.

## Changelog

:cl:
fix: Space dragons' eyes, teeths and claws (and shark form's underside)
are once again of a different color than the rest of the body.
/:cl:

---------

Co-authored-by: SmArtKar <44720187+SmArtKar@users.noreply.github.com>