Fix runtime when simple mobs are exposed to reagents (#89483)
## About The Pull Request
- Fixes #89485

There was a runtime that would occur when most simple mobs are exposed
to a reagent. This was due to the `reagents` holder being null. The fix
is just to do a simple null check.

## Why It's Good For The Game
No more runtimes when it's raining m'kay.

## Changelog
:cl:
fix: Fix runtime when simple mobs are exposed to reagents
/:cl: