You can no longer use cans of bait on fishing rods while fishing with them (#89528)
## About The Pull Request
Can of bait can be used on a fishing rod directly to quickly put a bait
into it's bait slot. However the proc that retrieves the bait from the
can is called before the code actually checks if you can put anything
into the fishing rod. This causes the can to go on cooldown and
decrement its bait count without it actually doing anything with the
bait.

This moves the bait retrieval from attackby into use_slot, after it
checks if it can use the item.

This also allows the can of bait to be used on the bait slot inside the
rod's UI, which works the same way as clicking on the rod.
## Why It's Good For The Game
Fixes #88216 and makes the use of fishing rod's UI and direct clicking
more consistent.
## Changelog
:cl:
fix: Can of fishing bait no longer evaporates its bait when used on an
active fishing rod
fix: Can of fishing bait's quick refill can be used with the fishing
rod's UI too
/:cl: