Trash contains material (#89667)
## About The Pull Request
Adds small amounts of material to most trash items. Generally, these are
about a 5th of a sheet's worth of either plastic or iron.

NOTE: The trash recycler on the station is 62.5% efficient using default
parts. If you use it you are going to get a sheet after around 8 trash
items, not 5.
## Why It's Good For The Game
Previously the only reward for recycling these items was removing them.
Now you can keep that desk area tidy and help out the station by
providing a small trickle of plastic and iron, just from flushing your
garbage.

The iron isn't a big deal but this is a way for any department to
produce small amounts of plastic, service in particular. It's
labour-intensive and the yields are small, but now clowns can pick up
trash and flush it for materials to make their balloons for example.
Players can even buy food items at the vendors for the sole purpose of
making them into trash and making the trash into plastic, if the
situation is dire.
## Changelog
:cl:
balance: Trash items now contain small amounts of material.
/:cl:

---------

Co-authored-by: tgstation-ci[bot] <179393467+tgstation-ci[bot]@users.noreply.github.com>